Fix 2.1 regression for css-in-js
